protocol fix
removed location.protocol: you can use the relative URL //openepaperlink.eu/getupdate/?url= without specifying the protocol (http: or https:) in JavaScript when using the fetch() function. This is because the browser will automatically use the same protocol (HTTP or HTTPS) as the current page.
Also, reverted the http:// url in line 287 because that's an url that is called by the esp32, not the browser.